requirements.txt
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1618968359944/work
async-generator==1.10
attrs==21.4.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.11.1
category-encoders==2.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
cycler==0.11.0
debugpy @ file:///opt/concourse/worker/volumes/live/e17abc9f-08a9-4dca-4e9b-e6d24bde47ce/volume/debugpy_1637091828617/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1646044401614/work
Flask==2.1.1
fonttools==4.32.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1649684468949/work/dist/ipykernel-6.13.0-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1648413602721/work
itsdangerous==2.1.2
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1649067146773/work
Jinja2==3.1.1
joblib==1.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1649340884673/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1645024290809/work
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
numpy==1.22.3
outcome==1.1.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
patsy==0.5.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.1.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1649130487073/work
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1648873031030/work
psycopg2-binary==2.9.3
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
pyOpenSSL==22.0.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1649603503565/work
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.1
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1635877502229/work
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1644872665635/work
statsmodels==0.13.2
threadpoolctl==3.1.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1648827543297/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webdriver-manager==3.5.4
Werkzeug==2.1.1
wsproto==1.1.0
xgboost==1.6.0
zipp==3.8.0
